bracelet/requirements.txt
cryptography
PyJWT
pyOpenSSL
rfc3339
paho-mqtt